[MIG] l10n_th_account_tax_report: Migration to 16.0
