+++ l10n_th_account_tax_report/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     "name": "Thai Localization - VAT and Withholding Tax Reports",
-    "version": "15.0.1.1.0",
+    "version": "16.0.1.0.0",
     "author": "Ecosoft, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-thailand",
     "license": "AGPL-3",
@@ -21,9 +21,9 @@
         "data/report_data.xml",
         "reports/tax_report.xml",
         "reports/tax_report_rd.xml",
-        "reports/report_wht_qweb.xml",
-        "reports/report_wht_qweb_rd.xml",
-        "reports/report_wht_text.xml",
+        "reports/wht_report.xml",
+        "reports/wht_report_rd.xml",
+        "reports/wht_report_text.xml",
         "wizard/tax_report_wizard_view.xml",
         "wizard/withholding_tax_report_wizard_view.xml",
         "views/res_config_settings_views.xml",
@@ -32,8 +32,6 @@
     "assets": {
         "web.assets_backend": [
             "l10n_th_account_tax_report/static/src/scss/style_report.scss",
-            "l10n_th_account_tax_report/static/src/js/l10n_th_account_tax_report_backend.js",
-            "l10n_th_account_tax_report/static/src/js/withholding_tax_report_widgets.js",
         ],
         "web.report_assets_common": [
             "l10n_th_account_tax_report/static/src/scss/style_report.scss",
