+++ l10n_th_account_tax_report/data/paper_format.xml
@@ -23,15 +23,15 @@
         <field name="page_height">0</field>
         <field name="page_width">0</field>
         <field name="orientation">Landscape</field>
-        <field name="margin_top">45</field>
+        <field name="margin_top">50</field>
         <field name="margin_bottom">5</field>
         <field name="margin_left">5</field>
         <field name="margin_right">5</field>
         <field name="header_line" eval="False" />
-        <field name="header_spacing">40</field>
+        <field name="header_spacing">45</field>
         <field name="dpi">90</field>
     </record>
-    <record id="paperformat_wht_rd_report" model="report.paperformat">
+    <record id="paperformat_rd_wht_report" model="report.paperformat">
         <field name="name">WHT RD Report A4</field>
         <field name="default" eval="True" />
         <field name="format">A4</field>
