+++ l10n_th_account_tax_report/models/__init__.py
@@ -2,3 +2,4 @@
 
 from . import res_company
 from . import res_config_settings
+from . import ir_actions_report
