+++ l10n_th_account_tax_report/reports/__init__.py
@@ -1,7 +1,4 @@
-# Copyright 2019 Ecosoft Co., Ltd (https://ecosoft.co.th)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 
-from . import tax_report
 from . import tax_report_xlsx
-from . import report_wht_xlsx
-from . import report_withholding_tax
+from . import wht_report_xlsx
