+++ l10n_th_account_tax_report/reports/wht_report.xml
@@ -1,5 +1,16 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
+    <!-- Template Withholding Tax Report -->
+    <template id="report_withholding_tax">
+        <t t-call="web.html_container">
+            <t t-foreach="docs" t-as="o">
+                <t t-call="l10n_th_account_tax_report.internal_layout">
+                    <t t-call="l10n_th_account_tax_report.report_wht" />
+                </t>
+            </t>
+        </t>
+    </template>
+
     <!-- Header -->
     <template id="header_report_withholding_tax">
         <div class="act_as_table data_table">
@@ -133,7 +144,7 @@
             <div class="act_as_cell" />
         </div>
     </template>
-    <template id="report_withholding_tax">
+    <template id="report_wht">
         <t t-set="title">Withholding Tax Report</t>
         <div class="page">
             <div>
@@ -172,19 +183,10 @@
             </div>
         </div>
     </template>
-    <template id="report_withholding_tax_html">
+    <!-- <template id="report_withholding_tax_html">
         <div class="o_withholding_tax_report_page">
             <t t-call="l10n_th_account_tax_report.report_buttons" />
             <t t-call="l10n_th_account_tax_report.report_withholding_tax" />
         </div>
-    </template>
-    <template id="report_withholding_tax_qweb">
-        <t t-call="web.html_container">
-            <t t-foreach="docs" t-as="o">
-                <t t-call="l10n_th_account_tax_report.internal_layout">
-                    <t t-call="l10n_th_account_tax_report.report_withholding_tax" />
-                </t>
-            </t>
-        </t>
-    </template>
+    </template> -->
 </odoo>
