+++ l10n_th_account_tax_report/tests/__init__.py
@@ -1,4 +1,3 @@
-# Copyright 2021 Ecosoft Co., Ltd (https://ecosoft.co.th)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 
 from . import test_tax_report
